Simplify loops over receivers and components
Avoid allocating new arrays for receiver locations and just rely on
a generator over the receiver_list, saving some memory. Remove the
`recievers_indices` from the numba functions and just pass slices of the
sensitivity matrix corresponding to the rows that should be filled based
on the receivers and their components. Expand some docstrings.